requirements.txt
cloudpickle==0.4.0
exifread==2.1.2
fpdf2==2.1.0
joblib==0.14.1
matplotlib
networkx==1.11
numpy
Pillow>=8.1.1
pyproj>=1.9.5.1
pytest==3.0.7
python-dateutil>=2.7
pyyaml==5.4
scipy
Sphinx==3.4.3
six
xmltodict==0.10.2
beautifulsoup4==4.9.1
lxml==4.5.1
wheel
